Renamed Steganography to SteganographyUtils. Moved Utility classes to separate package.
